pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-sync"
version = "1.0.0"
description = "Production-ready CLI for synchronizing semantic models between Snowflake and Microsoft Fabric"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Platform Engineering Team", email = "platform@company.com"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["snowflake", "fabric", "semantic-model", "sync", "cli"]

dependencies = [
    "click>=8.1.0",
    "snowflake-connector-python>=3.0.0",
    "msal>=1.24.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
semantic-sync = "semantic_sync.main:cli"

[project.urls]
Homepage = "https://github.com/company/semantic-sync"
Documentation = "https://github.com/company/semantic-sync#readme"
Repository = "https://github.com/company/semantic-sync.git"
Issues = "https://github.com/company/semantic-sync/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["semantic_sync*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=semantic_sync --cov-report=term-missing"
